move APIFnCloseBundle to action_bundleclose.go
